feat(Driver): add new driver class
It acts as a mediator between DiscoveryService which can return various
endpoints and SessionPool which uses whatever endpoint is currently
ready and creates or gives a ready session.

Also implement:
* naive round-robin endpoint discovery method
* periodic endpoint discovery by DiscoveryService
* periodic session keep alive by SessionPool

Since both DiscoveryService and SessionPool now start periodic tasks,
they have to be explicitly destroyed, so that periodic tasks are
stopped and process terminates.