feat: process 'session-close' server hint from trailing metadata
To do so, first capture trailing metadata in `AuthenticatedService` base
class, using the request object as key and weak map data structure to
avoid memory leaks. Then extract that metadata in every `Session` method
and if the 'session-close' hint is found, mark that session as closing
one. That will result in deleting such a session once it is returned to
the session pool.

Also make it possible to access the trailing metadata inside user's code
while calling the `executeQuery` method, by extending the
`ExecuteQuerySettings` object with `onResponseMetadata` callback.