feat: introduce new type aliases: Types.BYTES & Types.TEXT
First one properly handles byte-array data: it converts string values to
Buffer before sending data to YDB and do not implicitly transform values
received from YDB to a string via any encoding, leaving this choice to
the client. Doing so allows to keep the old behavior for Types.STRING
type and avoid breaking the conversion of STRING values already stored
in YDB.

Types.TEXT behaves the same way as Types.UTF8 and was added just to
mirror a new type alias in YDB.

What was wrong with Types.STRING before? The most basic layer of ydb-sdk
(bundle.js got from protobufs) works with STRING and YSON values as
Node.js Buffer type. To handle these values properly, the same
transformation to/from Buffer should be applied on top of basic layer:
1. before sending the data to YDB, from string to Buffer (if string was
   used);
2. upon receiving data from YDB, from Buffer to string, in case the
   client expects the data to be string.

To maintain backwards compatibility (1) is not done for Types.STRING and
Types.YSON, so that the format of data already in YDB does not change.
The type of transformation (2) Types.STRING and Types.YSON is utf8, as
before.