fix: certificates path for bundling compilation
This is a poor man's fix for bundling compilers like esbuild\webpack
that require paths to be relative to the entry point.
Certs folder is moved into cjs\esm dists so we can utilize universal
path './certs' that would work for default tsc compilation and would
allow copying the certs folder to whatever path bundler is gonna output
compiled entry point.
Some sensible error communication is also added
This commit is not marked as breaking because if certs folder had not
been found when grpcs is present in previous implementation
the connection would not succeed and\or "File not found" error threw